bugs fixed

+++ src/main/java/by/itstep/application/repository/StudentRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
     Optional<Student> findByUser(User user);
 }
